Do some minor refactoring
